refactor(mocks/slice): migrate to typescript
